pyproject.toml
[project]
name = "build-with-ai-sessions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exa-py>=1.14.6",
    "firecrawl>=2.7.1",
    "instructor>=1.8.3",
    "logfire[httpx]>=3.19.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "openai-agents>=0.0.17",
    "openai>=1.86.0",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-instrumentation-httpx>=0.55b1",
    "opentelemetry-sdk>=1.34.1",
    "pillow>=11.2.1",
    "pydantic>=2.11.6",
    "pydantic-ai[logfire]>=0.2.18",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
    "mcp-agent>=0.1.3",
    "mcp>=1.9.4",
    "a2a-sdk>=0.2.7",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "jupyterlab-rise>=0.43.1",
    "rise>=5.7.1",
    "ruff>=0.11.13",
    "watchdog>=6.0.0",
]
